🚀 release: QANTO Testnet Mainnet Alpha (v1.0.0-alpha)
QANTO Layer-0 Protocol reaches the Testnet Mainnet Alpha baseline.

Key milestones in this release:
- Post-quantum cryptography: CRYSTALS-Dilithium + Kyber KEM + SPHINCS+
- Hybrid PoW+DPoS consensus with SAGA AI governance
- DAG-based parallel processing targeting 25M+ TPS / 50ms finality
- IBC-style cross-chain interoperability
- Infinite Strata mining with VDF proofs
- Deterministic BIP39/HKDF key derivation for wallets
- Canonical bincode serialization for transaction integrity
- Genesis loader driven by config/genesis.json (no hardcoded UTXOs)
- Consensus lock (tokio::sync::Mutex) guards double-spend race conditions
- Tauri desktop wallet with real-time chain-native balance streaming
- Kubernetes/Helm/Docker production deployment manifests
- CI/CD workflows for multi-platform wallet release (Win/Mac/Linux)

Compilation: cargo check --workspace --all-targets passes with Exit code 0.
Artifacts cleaned: .dbg, logs, .testnet, keystore.json, wallet.key, p2p_identity.key.
.gitignore hardened for build artifacts, cargo targets, and runtime secrets.

This commit declares QANTO ready for Public Testnet ignition.

+++ Protocol_Core/myblockchain/benches/main.rs
@@ -41,7 +41,7 @@ fn qanhash_op_benchmark(c: &mut Criterion) {
                 res = qanhash::hash(
                     black_box(&header_hash),
                     black_box(start_nonce + i),
-                    black_box(i as u64),
+                    black_box(i),
                 );
             }
             res
